переностит функцию-цикл генерации объектов в data, переносит её вызов в advertisment.js
