+++ js/advertisment.js
@@ -1,0 +1,3 @@
+import {generateObjects} from './data.js';
+
+generateObjects(1);
